fix: guard central google connect flow
