+++ SClass-Infrastructure/src/main/kotlin/com/sclass/infrastructure/calendar/CentralGoogleCalendarClient.kt
@@ -1,10 +1,12 @@
 package com.sclass.infrastructure.calendar
 
+import com.sclass.common.exception.GoogleCalendarUnauthorizedException
 import com.sclass.infrastructure.calendar.dto.GoogleCalendarEventCreateCommand
 import com.sclass.infrastructure.calendar.dto.GoogleCalendarEventResult
 import com.sclass.infrastructure.oauth.client.CentralGoogleAuthorizationCodeClient
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.stereotype.Component
+import org.springframework.web.reactive.function.client.WebClientResponseException
 
 @Component
 @ConditionalOnProperty(prefix = "google.calendar.central", name = ["enabled"], havingValue = "true")
@@ -18,10 +20,23 @@ class CentralGoogleCalendarClient(
         command: GoogleCalendarEventCreateCommand,
     ): GoogleCalendarEventResult {
         val accessToken = authorizationCodeClient.refreshAccessToken(refreshToken)
-        return googleCalendarClient.createMeetEventWithAccessToken(
-            command = command,
-            accessToken = accessToken,
-            calendarId = properties.calendarId,
-        )
+        return try {
+            googleCalendarClient.createMeetEventWithAccessToken(
+                command = command,
+                accessToken = accessToken,
+                calendarId = properties.calendarId,
+            )
+        } catch (e: WebClientResponseException) {
+            if (e.isAuthorizationFailure()) throw GoogleCalendarUnauthorizedException()
+            throw e
+        }
+    }
+
+    private fun WebClientResponseException.isAuthorizationFailure(): Boolean =
+        statusCode.value() == UNAUTHORIZED_STATUS || statusCode.value() == FORBIDDEN_STATUS
+
+    private companion object {
+        const val UNAUTHORIZED_STATUS = 401
+        const val FORBIDDEN_STATUS = 403
     }
 }
