+++ SClass-Infrastructure/src/test/kotlin/com/sclass/infrastructure/calendar/CentralGoogleCalendarClientTest.kt
@@ -1,5 +1,6 @@
 package com.sclass.infrastructure.calendar
 
+import com.sclass.common.exception.GoogleCalendarUnauthorizedException
 import com.sclass.infrastructure.calendar.dto.GoogleCalendarEventCreateCommand
 import com.sclass.infrastructure.calendar.dto.GoogleCalendarEventResult
 import com.sclass.infrastructure.oauth.client.CentralGoogleAuthorizationCodeClient
@@ -8,6 +9,10 @@ import io.mockk.mockk
 import io.mockk.verify
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.assertThrows
+import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
+import org.springframework.web.reactive.function.client.WebClientResponseException
 import java.time.ZoneId
 import java.time.ZonedDateTime
 
@@ -27,13 +32,7 @@ class CentralGoogleCalendarClientTest {
 
     @Test
     fun `central refresh token으로 access token을 갱신해 중앙 캘린더에 Meet 이벤트를 생성한다`() {
-        val command =
-            GoogleCalendarEventCreateCommand(
-                summary = "수학 1회차",
-                startAt = ZonedDateTime.of(2026, 4, 26, 14, 0, 0, 0, ZoneId.of("Asia/Seoul")),
-                endAt = ZonedDateTime.of(2026, 4, 26, 15, 0, 0, 0, ZoneId.of("Asia/Seoul")),
-                attendeeEmails = listOf("teacher@example.com", "student@example.com"),
-            )
+        val command = command()
         val expected = GoogleCalendarEventResult(eventId = "event-id", meetJoinUrl = "https://meet.google.com/abc-defg-hij")
 
         every { authorizationCodeClient.refreshAccessToken("central-refresh-token") } returns "central-access-token"
@@ -61,4 +60,39 @@ class CentralGoogleCalendarClientTest {
             )
         }
     }
+
+    @Test
+    fun `중앙 캘린더 이벤트 생성에서 Google 인증 실패가 발생하면 도메인 예외로 변환한다`() {
+        val command = command()
+        every { authorizationCodeClient.refreshAccessToken("central-refresh-token") } returns "central-access-token"
+        every {
+            googleCalendarClient.createMeetEventWithAccessToken(
+                command = command,
+                accessToken = "central-access-token",
+                calendarId = "primary",
+            )
+        } throws
+            WebClientResponseException.create(
+                HttpStatus.UNAUTHORIZED.value(),
+                "Unauthorized",
+                HttpHeaders.EMPTY,
+                ByteArray(0),
+                null,
+            )
+
+        assertThrows<GoogleCalendarUnauthorizedException> {
+            client.createMeetEventWithRefreshToken(
+                refreshToken = "central-refresh-token",
+                command = command,
+            )
+        }
+    }
+
+    private fun command() =
+        GoogleCalendarEventCreateCommand(
+            summary = "수학 1회차",
+            startAt = ZonedDateTime.of(2026, 4, 26, 14, 0, 0, 0, ZoneId.of("Asia/Seoul")),
+            endAt = ZonedDateTime.of(2026, 4, 26, 15, 0, 0, 0, ZoneId.of("Asia/Seoul")),
+            attendeeEmails = listOf("teacher@example.com", "student@example.com"),
+        )
 }
